requirements.txt
# Core Dependencies
streamlit>=1.29.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.5.0

# LLM Providers
openai>=1.7.0
anthropic>=0.8.0
cohere>=4.40
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langgraph>=0.0.20
langchain-core>=0.1.0

# Vector Databases
chromadb>=0.4.22
faiss-cpu>=1.7.4
pinecone-client>=3.0.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.11.0
tiktoken>=0.5.2

# Embeddings
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0

# Evaluation
ragas>=0.1.0
datasets>=2.16.0

# Monitoring & Logging
loguru>=0.7.2

# Utilities
numpy>=1.24.0
pandas>=2.1.0
requests>=2.31.0
tavily-python>=0.3.0

# AWS Deployment (optional)
boto3>=1.34.0
sagemaker>=2.200.0

# Azure Deployment (optional)
azure-identity>=1.15.0
azure-mgmt-containerinstance>=10.1.0
azure-mgmt-web>=7.2.0

# HuggingFace Deployment (optional)
huggingface-hub>=0.20.0
